회원 예약/예약취소 작업 끝!

+++ src/main/java/kr/human/second/vo/ReservationVO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class ReservationVO {
 	private String id;
-	private long idx;
+	private int idx;
 }
